+++ src/main/java/org/proshin/finapi/account/out/SepaRequestingResponse.java
@@ -23,6 +23,8 @@ public interface SepaRequestingResponse {
 
     Optional<String> warnMessage();
 
+    Long paymentId();
+
     Optional<String> challengeMessage();
 
     Optional<String> answerFieldLabel();