+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#130](https://github.com/proshin-roman/finapi-java-client/issues/130) - The library now supports finAPI v1.64.0
 
 ## [0.1.63] - 2019-02-16
 ### Added